Define a template for notification e-mails
Makes it possible to define a template file that is used to create the body of
e-mails sent by the Notification Mail client. The path to the template file is
specified via the argument `mail_template_path`. Mustache has been chosen as
the template engine as it well documented and introduces no other dependencies.

Mails can now not only be sent as plain text, but also formatted as HTML. This
behaviour needs to be turned on by specifying the flag `mail_as_html`.

Also specs have been added for `MailClient`. The test dependency `akka-testkit`
has been added in order to make testing of actors possible.